+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1065](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1065))
 - `opentelemetry-instrumentation-redis` now instruments asynchronous Redis clients, if the installed redis-py includes async support (>=4.2.0).
   ([#1076](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1076))
+- `opentelemetry-instrumentation-boto3sqs` added AWS's SQS instrumentation.
+  ([#1081](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1081))
 
 ## [1.11.1-0.30b1](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v1.11.1-0.30b1) - 2022-04-21